[SNAP-1140] Performance issues seen in query re-routing (#119)
- Re-routed prepared statements are never marked as invalid so they are always re-prepared and multiple times
  in one round (because reprepare gets invoked from multiple places). Mark as valid and also increment version
  counter which is used by DRDA server to invalidate client side descriptors in case it gets re-prepared due
  to a genuine reason.
- Disable nagle's algorithm on the server-side DRDA socket. Otherwise queries returning small results can
  block for quite long time (e.g. ~40ms delay seen for a query returning single row with two columns which
      was tracked down to the nagle's algorithm on server-side socket); client socket already disables it
- Move out pattern matching for INSERT/PUT...SELECT for re-routing to be pre-compiled static. Also changed
  the STREAM pattern to be a regex rather than string match which is buggy. In addition removed the
  toUpperCase() of source which is very inefficient way of matching and instead use CASE_INSENSITIVE regex.
- Changed some instances of System.currentTimeInMillis() with nanoTime() spotted while reviewing code
  which is much more efficient and also much more accurate for measuring intervals. Changed SessionVTI
  interval type to be DOUBLE rather than BIGINT (latter was useless since time is shown in seconds).

+++ gemfirexd/core/src/main/java/com/pivotal/gemfirexd/internal/engine/diag/SessionsVTI.java
@@ -165,7 +165,7 @@ public class SessionsVTI extends GfxdVTITemplate {
           CURRENT_STATEMENT_STATUS, Types.VARCHAR, true, 32),
           
       EmbedResultSetMetaData.getResultColumnDescriptor(
-          CURRENT_STATEMENT_ELAPSED_TIME, Types.BIGINT, true),
+          CURRENT_STATEMENT_ELAPSED_TIME, Types.DOUBLE, true),
           
       EmbedResultSetMetaData.getResultColumnDescriptor(
           CURRENT_STATEMENT_ACCESS_FREQUENCY, Types.BIGINT, true),
@@ -331,7 +331,7 @@ public class SessionsVTI extends GfxdVTITemplate {
 
       public String currentStatementStatus;
       
-      public long currentStatementElapsedTime;
+      public double currentStatementElapsedTime;
       
       public long currentStatementAccessFrequency;
